requirements.txt
flask
newspaper3k
nltk
scikit-learn
lxml_html_clean